Modify counter assignments to better utilize registers
Previously, the counters used to track register numbers were incremented
by a constant value in each loop iteration; this yielded inefficient use
of registers under the constraint of limited registers. The code that
keeps track of counters has been modified so that a minimal amount of
registers can be used.